check if file exists then remove

+++ aimodelshare/api.py
@@ -154,8 +154,10 @@ def create_prediction_api(my_credentials, model_filepath, unique_model_id, model
     except Exception as e:
         print(e)
     
-    os.remove(os.path.join(temp_dir,'main.py'))
-
+    if os.path.exists(os.path.join(temp_dir,'main.py')):
+      os.remove(os.path.join(temp_dir,'main.py'))
+    else:
+      pass     
 # Upload model eval lambda function zipfile to user's model file folder on s3
     if categorical == 'TRUE':
             data = pkg_resources.read_text(main, 'eval_classification.txt')
